refactor common ref #7
